backend/requirements.txt
requests>=2.32
feedparser>=6.0
schedule>=1.2
python-dotenv>=1.0
